Simplified code and fixed issues
* Added dry run option to Merger
* Changed behavior when actual merge differs from dry run to just output a warning to console
* Fixed KeeShare conflicting with merge operations in the middle of a merge

+++ src/core/Merger.cpp
@@ -49,6 +49,17 @@ Merger::Change::Change(QString details)
 {
 }
 
+bool Merger::Change::operator==(const Merger::Change& other) const
+{
+    return m_type == other.m_type && m_group == other.m_group && m_title == other.m_title && m_uuid == other.m_uuid
+           && m_details == other.m_details;
+}
+
+bool Merger::Change::operator!=(const Merger::Change& other) const
+{
+    return !(*this == other);
+}
+
 Merger::Change::Type Merger::Change::type() const
 {
     return m_type;
@@ -111,12 +122,6 @@ QString Merger::Change::toString() const
     return result;
 }
 
-bool operator==(const Merger::Change& lhs, const Merger::Change& rhs)
-{
-    return lhs.type() == rhs.type() && lhs.group() == rhs.group() && lhs.title() == rhs.title()
-           && lhs.uuid() == rhs.uuid() && lhs.details() == rhs.details();
-}
-
 Merger::Merger(const Database* sourceDb, Database* targetDb)
     : m_mode(Group::Default)
 {
@@ -160,8 +165,9 @@ void Merger::setSkipDatabaseCustomData(bool state)
     m_skipCustomData = state;
 }
 
-Merger::ChangeList Merger::merge()
+Merger::ChangeList Merger::merge(bool dryRun)
 {
+    m_dryRun = dryRun;
     // Order of merge steps is important - it is possible that we
     // create some items before deleting them afterwards
     ChangeList changes;
@@ -170,9 +176,10 @@ Merger::ChangeList Merger::merge()
     changes << mergeMetadata(m_context);
 
     // At this point we have a list of changes we may want to show the user
-    if (!changes.isEmpty()) {
+    if (!changes.isEmpty() && !dryRun) {
         m_context.m_targetDb->markAsModified();
     }
+    m_dryRun = false;
     return changes;
 }
 
@@ -184,10 +191,12 @@ Merger::ChangeList Merger::mergeGroup(const MergeContext& context)
     for (Entry* sourceEntry : sourceEntries) {
         Entry* targetEntry = context.m_targetRootGroup->findEntryByUuid(sourceEntry->uuid());
         if (!targetEntry) {
-            changes << Change(Change::Type::Added, *sourceEntry);
             // This entry does not exist at all. Create it.
-            targetEntry = sourceEntry->clone(Entry::CloneIncludeHistory);
-            moveEntry(targetEntry, context.m_targetGroup);
+            changes << Change(Change::Type::Added, *sourceEntry);
+            if (!m_dryRun) {
+                targetEntry = sourceEntry->clone(Entry::CloneIncludeHistory);
+                moveEntry(targetEntry, context.m_targetGroup);
+            }
         } else {
             // Entry is already present in the database. Update it.
             const bool locationChanged =
@@ -195,36 +204,45 @@ Merger::ChangeList Merger::mergeGroup(const MergeContext& context)
             if (locationChanged && targetEntry->group() != context.m_targetGroup) {
                 changes << Change(
                     Change::Type::Moved, *sourceEntry, tr("Previous location: %1").arg(targetEntry->group()->name()));
-                moveEntry(targetEntry, context.m_targetGroup);
+                if (!m_dryRun) {
+                    moveEntry(targetEntry, context.m_targetGroup);
+                }
             }
             changes << resolveEntryConflict(context, sourceEntry, targetEntry);
         }
     }
 
-    // merge groups recursively
+    // merge child groups recursively
     const QList<Group*> sourceChildGroups = context.m_sourceGroup->children();
     for (Group* sourceChildGroup : sourceChildGroups) {
+        bool groupCreated = false;
         Group* targetChildGroup = context.m_targetRootGroup->findGroupByUuid(sourceChildGroup->uuid());
         if (!targetChildGroup) {
             changes << Change(
                 Change::Type::Added,
                 *sourceChildGroup,
                 tr("Number of entries in group: %1").arg(QString::number(sourceChildGroup->entries().size())));
+            // Create the target group, it will be cleaned up later if in dry run mode
             targetChildGroup = sourceChildGroup->clone(Entry::CloneNoFlags, Group::CloneNoFlags);
-            moveGroup(targetChildGroup, context.m_targetGroup);
-            TimeInfo timeinfo = targetChildGroup->timeInfo();
-            timeinfo.setLocationChanged(sourceChildGroup->timeInfo().locationChanged());
-            targetChildGroup->setTimeInfo(timeinfo);
+            groupCreated = true;
+            if (!m_dryRun) {
+                moveGroup(targetChildGroup, context.m_targetGroup);
+                TimeInfo timeinfo = targetChildGroup->timeInfo();
+                timeinfo.setLocationChanged(sourceChildGroup->timeInfo().locationChanged());
+                targetChildGroup->setTimeInfo(timeinfo);
+            }
         } else {
             bool locationChanged =
                 targetChildGroup->timeInfo().locationChanged() < sourceChildGroup->timeInfo().locationChanged();
             if (locationChanged && targetChildGroup->parent() != context.m_targetGroup) {
                 changes << Change(
                     Change::Type::Moved, *sourceChildGroup, tr("Previous location: %1").arg(targetChildGroup->name()));
-                moveGroup(targetChildGroup, context.m_targetGroup);
-                TimeInfo timeinfo = targetChildGroup->timeInfo();
-                timeinfo.setLocationChanged(sourceChildGroup->timeInfo().locationChanged());
-                targetChildGroup->setTimeInfo(timeinfo);
+                if (!m_dryRun) {
+                    moveGroup(targetChildGroup, context.m_targetGroup);
+                    TimeInfo timeinfo = targetChildGroup->timeInfo();
+                    timeinfo.setLocationChanged(sourceChildGroup->timeInfo().locationChanged());
+                    targetChildGroup->setTimeInfo(timeinfo);
+                }
             }
             changes << resolveGroupConflict(context, sourceChildGroup, targetChildGroup);
         }
@@ -235,6 +253,10 @@ Merger::ChangeList Merger::mergeGroup(const MergeContext& context)
                                 sourceChildGroup,
                                 targetChildGroup};
         changes << mergeGroup(subcontext);
+        // Cleanup the temporary target group structure
+        if (m_dryRun && groupCreated) {
+            delete subcontext.m_targetGroup;
+        }
     }
     return changes;
 }
@@ -251,14 +273,15 @@ Merger::resolveGroupConflict(const MergeContext& context, const Group* sourceChi
     // only if the other group is newer, update the existing one.
     if (timeExisting < timeOther) {
         QStringList modifications;
-        auto updateIfNecessary = [&modifications](const auto& targetValue,
-                                                  const auto& sourceValue,
-                                                  auto&& updateFunction,
-                                                  const QString& modification) {
+        auto updateIfNecessary = [&modifications, this](const auto& targetValue,
+                                                        const auto& sourceValue,
+                                                        auto&& updateFunction,
+                                                        const QString& modification) {
             if (targetValue != sourceValue) {
-                updateFunction(sourceValue);
-
                 modifications << modification;
+                if (!m_dryRun) {
+                    updateFunction(sourceValue);
+                }
                 return true;
             }
             return false;
@@ -307,6 +330,10 @@ Merger::resolveGroupConflict(const MergeContext& context, const Group* sourceChi
 
 void Merger::moveEntry(Entry* entry, Group* targetGroup)
 {
+    if (m_dryRun) {
+        return;
+    }
+
     Q_ASSERT(entry);
     Group* sourceGroup = entry->group();
     if (sourceGroup == targetGroup) {
@@ -336,6 +363,10 @@ void Merger::moveEntry(Entry* entry, Group* targetGroup)
 
 void Merger::moveGroup(Group* group, Group* targetGroup)
 {
+    if (m_dryRun) {
+        return;
+    }
+
     Q_ASSERT(group);
     Group* sourceGroup = group->parentGroup();
     if (sourceGroup == targetGroup) {
@@ -365,6 +396,10 @@ void Merger::moveGroup(Group* group, Group* targetGroup)
 
 void Merger::eraseEntry(Entry* entry)
 {
+    if (m_dryRun) {
+        return;
+    }
+
     Database* database = entry->database();
     // most simple method to remove an item from DeletedObjects :(
     const QList<DeletedObject> deletions = database->deletedObjects();
@@ -382,6 +417,10 @@ void Merger::eraseEntry(Entry* entry)
 
 void Merger::eraseGroup(Group* group)
 {
+    if (m_dryRun) {
+        return;
+    }
+
     Database* database = group->database();
     // most simple method to remove an item from DeletedObjects :(
     const QList<DeletedObject> deletions = database->deletedObjects();
@@ -418,10 +457,10 @@ Merger::ChangeList Merger::resolveEntryConflict_MergeHistories(const MergeContex
                qPrintable(targetEntry->title()),
                qPrintable(sourceEntry->title()),
                qPrintable(currentGroup->name()));
-        mergeHistory(targetEntry, clonedEntry, mergeMethod, maxItems);
         changes << Change(Change::Type::Modified,
                           *targetEntry,
                           tr("%1 (Add local modifications to new entry)").arg(differences.join(", ")));
+        mergeHistory(targetEntry, clonedEntry, mergeMethod, maxItems);
         eraseEntry(targetEntry);
         moveEntry(clonedEntry, currentGroup);
     } else {
@@ -445,8 +484,10 @@ Merger::resolveEntryConflict(const MergeContext& context, const Entry* sourceEnt
     // We need to cut off the milliseconds since the persistent format only supports times down to seconds
     // so when we import data from a remote source, it may represent the (or even some msec newer) data
     // which may be discarded due to higher runtime precision
-
-    Group::MergeMode mergeMode = m_mode == Group::Default ? context.m_targetGroup->mergeMode() : m_mode;
+    Group::MergeMode mergeMode = m_mode;
+    if (mergeMode == Group::Default && context.m_targetGroup) {
+        mergeMode = context.m_targetGroup->mergeMode();
+    }
     return resolveEntryConflict_MergeHistories(context, sourceEntry, targetEntry, mergeMode);
 }
 
@@ -469,11 +510,11 @@ bool Merger::mergeHistory(const Entry* sourceEntry,
         const QDateTime modificationTime = Clock::serialized(historyItem->timeInfo().lastModificationTime());
         if (merged.contains(modificationTime)
             && !merged[modificationTime]->equals(historyItem, CompareItemIgnoreMilliseconds)) {
-            ::qWarning("Inconsistent history entry of %s[%s] at %s contains conflicting changes - conflict resolution "
-                       "may lose data!",
-                       qPrintable(sourceEntry->title()),
-                       qPrintable(sourceEntry->uuidToHex()),
-                       qPrintable(modificationTime.toString("yyyy-MM-dd HH-mm-ss-zzz")));
+            qWarning("Inconsistent history entry of %s[%s] at %s contains conflicting changes - conflict resolution "
+                     "may lose data!",
+                     qPrintable(sourceEntry->title()),
+                     qPrintable(sourceEntry->uuidToHex()),
+                     qPrintable(modificationTime.toString("yyyy-MM-dd HH-mm-ss-zzz")));
         }
         merged[modificationTime] = historyItem->clone(Entry::CloneNoFlags);
     }
@@ -482,11 +523,10 @@ bool Merger::mergeHistory(const Entry* sourceEntry,
         const QDateTime modificationTime = Clock::serialized(historyItem->timeInfo().lastModificationTime());
         if (merged.contains(modificationTime)
             && !merged[modificationTime]->equals(historyItem, CompareItemIgnoreMilliseconds)) {
-            ::qWarning(
-                "History entry of %s[%s] at %s contains conflicting changes - conflict resolution may lose data!",
-                qPrintable(sourceEntry->title()),
-                qPrintable(sourceEntry->uuidToHex()),
-                qPrintable(modificationTime.toString("yyyy-MM-dd HH-mm-ss-zzz")));
+            qWarning("History entry of %s[%s] at %s contains conflicting changes - conflict resolution may lose data!",
+                     qPrintable(sourceEntry->title()),
+                     qPrintable(sourceEntry->uuidToHex()),
+                     qPrintable(modificationTime.toString("yyyy-MM-dd HH-mm-ss-zzz")));
         }
         if (preferRemote && merged.contains(modificationTime)) {
             // forcefully apply the remote history item
@@ -502,9 +542,9 @@ bool Merger::mergeHistory(const Entry* sourceEntry,
     if (targetModificationTime == sourceModificationTime
         && !targetEntry->equals(sourceEntry,
                                 CompareItemIgnoreMilliseconds | CompareItemIgnoreHistory | CompareItemIgnoreLocation)) {
-        ::qWarning("Entry of %s[%s] contains conflicting changes - conflict resolution may lose data!",
-                   qPrintable(sourceEntry->title()),
-                   qPrintable(sourceEntry->uuidToHex()));
+        qWarning("Entry of %s[%s] contains conflicting changes - conflict resolution may lose data!",
+                 qPrintable(sourceEntry->title()),
+                 qPrintable(sourceEntry->uuidToHex()));
     }
 
     if (targetModificationTime < sourceModificationTime) {
@@ -543,22 +583,24 @@ bool Merger::mergeHistory(const Entry* sourceEntry,
         qDeleteAll(updatedHistoryItems);
         return false;
     }
-    // We need to prevent any modification to the database since every change should be tracked either
-    // in a clone history item or in the Entry itself
-    const TimeInfo timeInfo = targetEntry->timeInfo();
-    const bool blockedSignals = targetEntry->blockSignals(true);
-    bool updateTimeInfo = targetEntry->canUpdateTimeinfo();
-    targetEntry->setUpdateTimeinfo(false);
-    targetEntry->removeHistoryItems(targetHistoryItems);
-    for (Entry* historyItem : merged) {
-        Q_ASSERT(!historyItem->parent());
-        targetEntry->addHistoryItem(historyItem);
+    if (!m_dryRun) {
+        // We need to prevent any modification to the database since every change should be tracked either
+        // in a clone history item or in the Entry itself
+        const TimeInfo timeInfo = targetEntry->timeInfo();
+        const bool blockedSignals = targetEntry->blockSignals(true);
+        bool updateTimeInfo = targetEntry->canUpdateTimeinfo();
+        targetEntry->setUpdateTimeinfo(false);
+        targetEntry->removeHistoryItems(targetHistoryItems);
+        for (Entry* historyItem : merged) {
+            Q_ASSERT(!historyItem->parent());
+            targetEntry->addHistoryItem(historyItem);
+        }
+        targetEntry->truncateHistory();
+        targetEntry->blockSignals(blockedSignals);
+        targetEntry->setUpdateTimeinfo(updateTimeInfo);
+        Q_ASSERT(timeInfo == targetEntry->timeInfo());
+        Q_UNUSED(timeInfo);
     }
-    targetEntry->truncateHistory();
-    targetEntry->blockSignals(blockedSignals);
-    targetEntry->setUpdateTimeinfo(updateTimeInfo);
-    Q_ASSERT(timeInfo == targetEntry->timeInfo());
-    Q_UNUSED(timeInfo);
     return true;
 }
 
@@ -582,7 +624,6 @@ Merger::ChangeList Merger::mergeDeletions(const MergeContext& context)
     for (const auto& object : (targetDeletions + sourceDeletions)) {
         if (!mergedDeletions.contains(object.uuid)) {
             mergedDeletions[object.uuid] = object;
-
             auto* entry = context.m_targetRootGroup->findEntryByUuid(object.uuid);
             if (entry) {
                 entries << entry;
@@ -614,8 +655,9 @@ Merger::ChangeList Merger::mergeDeletions(const MergeContext& context)
         } else {
             changes << Change(Change::Type::Deleted, *entry, tr("Implicit deletion (e.g. removal of parent group)"));
         }
-        // Entry is inserted into deletedObjects after deletions are processed
-        eraseEntry(entry);
+        if (!m_dryRun) {
+            eraseEntry(entry);
+        }
     }
 
     while (!groups.isEmpty()) {
@@ -640,13 +682,18 @@ Merger::ChangeList Merger::mergeDeletions(const MergeContext& context)
         } else {
             changes << Change(Change::Type::Deleted, *group, tr("Implicit deletion (e.g. removal of parent group)"));
         }
-        eraseGroup(group);
+        if (!m_dryRun) {
+            eraseGroup(group);
+        }
     }
     // Put every deletion to the earliest date of deletion
     if (deletions != context.m_targetDb->deletedObjects()) {
         changes << Change(Change::Type::Metadata, tr("Changed deleted objects"));
+        if (!m_dryRun) {
+            context.m_targetDb->setDeletedObjects(deletions);
+        }
     }
-    context.m_targetDb->setDeletedObjects(deletions);
+
     return changes;
 }
 
@@ -661,9 +708,11 @@ Merger::ChangeList Merger::mergeMetadata(const MergeContext& context)
 
     for (const auto& iconUuid : sourceMetadata->customIconsOrder()) {
         if (!targetMetadata->hasCustomIcon(iconUuid)) {
-            targetMetadata->addCustomIcon(iconUuid, sourceMetadata->customIcon(iconUuid));
             changes << Change(Change::Type::Metadata,
                               tr("Adding new icon %1").arg(QString::fromLatin1(iconUuid.toRfc4122().toHex())));
+            if (!m_dryRun) {
+                targetMetadata->addCustomIcon(iconUuid, sourceMetadata->customIcon(iconUuid));
+            }
         }
     }
 
@@ -686,8 +735,10 @@ Merger::ChangeList Merger::mergeMetadata(const MergeContext& context)
             // Do not remove protected custom data
             if (!sourceMetadata->customData()->contains(key) && !sourceMetadata->customData()->isProtected(key)) {
                 auto value = targetMetadata->customData()->value(key);
-                targetMetadata->customData()->remove(key);
                 changes << Change(Change::Type::Metadata, tr("Removed custom data %1 [%2]").arg(key, value));
+                if (!m_dryRun) {
+                    targetMetadata->customData()->remove(key);
+                }
             }
         }
 
@@ -702,8 +753,10 @@ Merger::ChangeList Merger::mergeMetadata(const MergeContext& context)
             auto targetValue = targetMetadata->customData()->value(key);
             // Merge only if the values are not the same.
             if (sourceValue != targetValue) {
-                targetMetadata->customData()->set(key, sourceValue);
                 changes << Change(Change::Type::Metadata, tr("Adding custom data %1 [%2]").arg(key, sourceValue));
+                if (!m_dryRun) {
+                    targetMetadata->customData()->set(key, sourceValue);
+                }
             }
         }
     }
